requirements.txt
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
fonttools==4.33.3
idna==3.3
joblib==1.1.0
kiwisolver==1.4.3
matplotlib==3.5.2
mypy==0.961
mypy-extensions==0.4.3
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
threadpoolctl==3.1.0
timm==0.6.2.dev0
tomli==2.0.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
yacs==0.1.8
